Remove boost::iostreams dependency and replace with Mio

+++ src/index/mempool_index.cpp
@@ -20,11 +20,11 @@ namespace blocksci {
         return (boost::filesystem::path{baseDirectory}/"block_index").native();
     }
     
-    std::string MempoolIndex::nthTxFilePath(const std::string &baseDirectory, size_t i) {
+    std::string MempoolIndex::nthTxFilePath(const std::string &baseDirectory, int64_t i) {
         return (boost::filesystem::path{baseDirectory}/(std::to_string(i) + "_tx")).native();
     }
     
-    std::string MempoolIndex::nthBlockFilePath(const std::string &baseDirectory, size_t i) {
+    std::string MempoolIndex::nthBlockFilePath(const std::string &baseDirectory, int64_t i) {
         return (boost::filesystem::path{baseDirectory}/(std::to_string(i) + "_block")).native();
     }
 }
